# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Category {
  id: string;
  name: string;
  slug: string;
  description: string;
  imageUrl: string;
  productCount: number;
}

export interface Product {
  id: string;
  slug: string;
  name: string;
  brand: string;
  category: string;
  categorySlug: string;
  description: string;
  price: number;
  originalPrice: number;
  discountPercent: number;
  rating: number;
  reviewCount: number;
  stock: number;
  featured: boolean;
  isNew: boolean;
  colors: string[];
  sizes: string[];
  imageUrl: string;
  gallery: string[];
  tags: string[];
}

export interface ProductList {
  items: Product[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export type SearchSuggestionType = typeof SearchSuggestionType[keyof typeof SearchSuggestionType];


export const SearchSuggestionType = {
  product: 'product',
  brand: 'brand',
  category: 'category',
} as const;

export interface SearchSuggestion {
  label: string;
  type: SearchSuggestionType;
  slug: string;
  imageUrl: string;
}

export interface CartItemInput {
  productId: string;
  /** @minimum 1 */
  quantity: number;
  color?: string;
  size?: string;
}

export interface CartItemUpdate {
  /** @minimum 0 */
  quantity: number;
}

export interface CartItem {
  id: string;
  product: Product;
  quantity: number;
  /** @nullable */
  color: string | null;
  /** @nullable */
  size: string | null;
  lineTotal: number;
}

export interface Cart {
  items: CartItem[];
  itemCount: number;
  subtotal: number;
  discount: number;
  delivery: number;
  tax: number;
  total: number;
}

export interface WishlistInput {
  productId: string;
}

export interface CouponValidationInput {
  code: string;
  subtotal: number;
}

export interface CouponValidation {
  valid: boolean;
  code: string;
  discount: number;
  message: string;
}

export type CheckoutInputDeliveryMethod = typeof CheckoutInputDeliveryMethod[keyof typeof CheckoutInputDeliveryMethod];


export const CheckoutInputDeliveryMethod = {
  standard: 'standard',
  express: 'express',
} as const;

export type CheckoutInputPaymentMethod = typeof CheckoutInputPaymentMethod[keyof typeof CheckoutInputPaymentMethod];


export const CheckoutInputPaymentMethod = {
  card: 'card',
  upi: 'upi',
  cod: 'cod',
} as const;

export interface AddressInput {
  fullName: string;
  phone: string;
  line1: string;
  city: string;
  state: string;
  postalCode: string;
  country: string;
}

export interface CheckoutInput {
  address: AddressInput;
  deliveryMethod: CheckoutInputDeliveryMethod;
  paymentMethod: CheckoutInputPaymentMethod;
  /** @nullable */
  couponCode: string | null;
}

export interface OrderItem {
  product: Product;
  quantity: number;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
  returned: 'returned',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: string;
  createdAt: string;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  items: OrderItem[];
  total: number;
  address: AddressInput;
}

export interface Account {
  id: string;
  name: string;
  email: string;
  initials: string;
  orderCount: number;
  wishlistCount: number;
  addresses: AddressInput[];
}

export type AdminSummarySalesItem = {
  label: string;
  value: number;
};

export interface AdminSummary {
  revenue: number;
  orders: number;
  customers: number;
  products: number;
  lowStock: number;
  sales: AdminSummarySalesItem[];
}

export interface AdminProductInput {
  name: string;
  brand: string;
  categorySlug: string;
  price: number;
  originalPrice: number;
  stock: number;
  imageUrl: string;
}

export interface AdminProductUpdate {
  id: string;
  name?: string;
  price?: number;
  originalPrice?: number;
  stock?: number;
  featured?: boolean;
}

export type ListProductsParams = {
search?: string;
category?: string;
sort?: ListProductsSort;
minPrice?: number;
maxPrice?: number;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 48
 */
pageSize?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  recommended: 'recommended',
  newest: 'newest',
  'price-asc': 'price-asc',
  'price-desc': 'price-desc',
  rating: 'rating',
  popular: 'popular',
} as const;

export type GetSearchSuggestionsParams = {
/**
 * @minLength 1
 */
q: string;
};