# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NEXORA commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: Search and filter products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [recommended, newest, price-asc, price-desc, rating, popular]
            default: recommended
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 48
            default: 24
      responses:
        "200":
          description: Product listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{slug}/related:
    get:
      operationId: listRelatedProducts
      tags: [catalog]
      summary: Get related products
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /search/suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [catalog]
      summary: Get product search suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Search suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchSuggestion"
...
  /admin/products:
    post:
      operationId: createAdminProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateAdminProduct
      tags: [admin]
      summary: Update a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Product:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        brand: { type: string }
        category: { type: string }
        categorySlug: { type: string }
        description: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        discountPercent: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        stock: { type: integer }
        featured: { type: boolean }
        isNew: { type: boolean }
        colors:
          type: array
          items: { type: string }
        sizes:
          type: array
          items: { type: string }
        imageUrl: { type: string }
        gallery:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
      required:
        [id, slug, name, brand, category, categorySlug, description, price,
         originalPrice, discountPercent, rating, reviewCount, stock, featured,
         isNew, colors, sizes, imageUrl, gallery, tags]
    ProductList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
      required: [items, total, page, pageSize, totalPages]
    SearchSuggestion:
      type: object
      properties:
        label: { type: string }
        type: { type: string, enum: [product, brand, category] }
        slug: { type: string }
        imageUrl: { type: string }
      required: [label, type, slug, imageUrl]
...
    AdminProductInput:
      type: object
      properties:
        name: { type: string }
        brand: { type: string }
        categorySlug: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        stock: { type: integer }
        imageUrl: { type: string }
      required: [name, brand, categorySlug, price, originalPrice, stock, imageUrl]
    AdminProductUpdate:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        stock: { type: integer }
        featured: { type: boolean }
      required: [id]