# lib/api-zod/src/generated/types/adminProductInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminProductInput {
  name: string;
  brand: string;
  categorySlug: string;
  price: number;
  originalPrice: number;
  stock: number;
  imageUrl: string;
}